// src/app/models/crypto.model.ts
export interface Crypto {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  price_change_percentage_24h: number;
  market_cap: number;
  total_volume: number;
  high_24h: number;
  low_24h: number;
}

export interface PortfolioItem {
  crypto: Crypto;
  quantity: number;
  purchasePrice: number;
}